Test to prove messing auth check
