+++ source/ArchivedMessages.IntegrationTests/ArchivedMessagesWithOwnedRestrictionTests.cs
@@ -25,7 +25,7 @@ using Xunit.Abstractions;
 namespace Energinet.DataHub.EDI.ArchivedMessages.IntegrationTests;
 
 [Collection(nameof(ArchivedMessagesCollection))]
-public class SearchMessagesWithOwnedRestrictionTests : IAsyncLifetime
+public class ArchivedMessagesWithOwnedRestrictionTests : IAsyncLifetime
 {
     private readonly IArchivedMessagesClient _sut;
     private readonly ArchivedMessagesFixture _fixture;
@@ -35,7 +35,7 @@ public class SearchMessagesWithOwnedRestrictionTests : IAsyncLifetime
         Restriction.Owned,
         ActorRole.EnergySupplier);
 
-    public SearchMessagesWithOwnedRestrictionTests(ArchivedMessagesFixture fixture, ITestOutputHelper testOutputHelper)
+    public ArchivedMessagesWithOwnedRestrictionTests(ArchivedMessagesFixture fixture, ITestOutputHelper testOutputHelper)
     {
         _fixture = fixture;
 
@@ -55,6 +55,22 @@ public class SearchMessagesWithOwnedRestrictionTests : IAsyncLifetime
     public Task DisposeAsync()
     {
         return Task.CompletedTask;
+    }
+
+    [Fact]
+    public async Task Given_ArchivedMessagesInStorage_When_GettingMessageToAnotherActor_Then_GetsNoMessage()
+    {
+        // Arrange
+        var notAuthActor = "9999999999999";
+        var archivedMessage = await _fixture.CreateArchivedMessageAsync(
+            receiverNumber: notAuthActor,
+            senderNumber: notAuthActor);
+
+        // Act
+        var result = await _sut.GetAsync(archivedMessage.Id, CancellationToken.None);
+
+        // Assert
+        result.Should().BeNull();
     }
 
     [Fact]
